refactor(importlinter): drop lyra.stt/lyra.tts contracts post-deletion
Remove lyra.stt and lyra.tts from shared-modules-independence and
shared-modules-upper-boundary contracts. Fix stale relative imports
in hub.py (TYPE_CHECKING) and tts_dispatch.py to point to
core/ports/{stt,tts}.
